llamaindex/requirements.txt
arxiv
fastapi
langchain
langchain-community
llama_index
ollama
pydantic
PyYAML
qdrant_client
llama-index-vector-stores-qdrant
llama-index-llms-ollama
llama-index-embeddings-langchain
tqdm
pypdf
uvicorn
fastembed
spacy
sentence-transformers
gradio==3.50.2
tools
frontend
pymupdf
fitz
#python -m spacy download zh_core_web_sm